Mostly done on the 1.21 port. Just needs a few fixes

+++ common/src/main/java/com/copycatsplus/copycats/foundation/copycat/model/assembly/quad/QuadUVRotate.java
@@ -29,8 +29,8 @@ public record QuadUVRotate(Direction face, float pivotU, float pivotV, float rot
                 double cos = Math.cos(Math.toRadians(rotation));
                 double sin = Math.sin(Math.toRadians(rotation));
 
-                quad.vertices.get(vertex).uv.u = sprite.getU(u * cos - v * sin + pivotU);
-                quad.vertices.get(vertex).uv.v = sprite.getV(u * sin + v * cos + pivotV);
+                quad.vertices.get(vertex).uv.u = sprite.getU((float) (u * cos - v * sin + pivotU));
+                quad.vertices.get(vertex).uv.v = sprite.getV((float) (u * sin + v * cos + pivotV));
             }
         }
         return true;
